+++ Subsidio/Subsidio/Region.cs
@@ -12,7 +12,7 @@ namespace BibliotecaNegocio
         //Crear objeto de la Bdd
         private EntitiesSubsidio bdd = new EntitiesSubsidio();
 
-        public decimal Id_Region { get; set; }
+        public int Id_Region { get; set; }
 
         public string Nombre { get; set; }
